Fix Chrome autocomplete incorrectly triggerring actions
When using a Stimulus action like

    <button data-action="keydown.meta+k@window->search#open">

in Google Chrome, clicking an input's autocomplete suggestion
unexpectedly triggers the controller action even though no key is
pressed.

It seems that Chrome dispatches an Event with type === "keydown", but it
is not an instance of KeyboardEvent. This happens regardless of the key
filter.

Closes #852